added layouts and controllers
