+++ src/main/java/com/frasinu/iss/view/controllers/SteeringComController.java
@@ -3,14 +3,13 @@ package com.frasinu.iss.view.controllers;
 import com.frasinu.iss.persistance.model.Author;
 import com.frasinu.iss.service.AuthorService;
 import com.frasinu.iss.service.ConferenceEditionService;
-import com.frasinu.iss.service.ConferenceService;
 import com.frasinu.iss.service.UserService;
 import com.frasinu.iss.service.service_requests.author.CreateAuthorRequest;
 import com.frasinu.iss.service.service_requests.conferenceedition.FindByConferenceEditionIdRequest;
 import com.frasinu.iss.service.service_requests.user.FindByIdRequest;
 import com.frasinu.iss.service.service_requests.user.FindIfUserIsAuthorRequest;
-import com.frasinu.iss.view.Screen;
 import com.frasinu.iss.view.FrasinuApplication;
+import com.frasinu.iss.view.Screen;
 import javafx.event.ActionEvent;
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
@@ -21,11 +20,10 @@ import java.util.HashMap;
 import java.util.Optional;
 
 /**
- * Created by Ericqw on 15.05.2017.
+ * Created by Ericqw on 22.05.2017.
  */
-@Controller(value = "ScheduleController")
-public class ScheduleController extends BaseController{
-    private ConferenceService conferencesService;
+@Controller(value = "SteeringComController")
+public class SteeringComController extends BaseController {
     private ConferenceEditionService conferenceEditionService;
     private UserService userService;
     private AuthorService authorService;
@@ -44,24 +42,12 @@ public class ScheduleController extends BaseController{
     public void setAuthorService(AuthorService authorService) {
         this.authorService=authorService;
     }
-
-    @Autowired
-    public void setConferenceService(ConferenceService conferencesService) {
-        this.conferencesService=conferencesService;
+    public void seeSchedule(ActionEvent ac){
+        FrasinuApplication.changeScreen(Screen.SCHEDULE,getData());
     }
-
-    public void goToConferences(ActionEvent actionEvent) {
-        FrasinuApplication.changeScreen(Screen.CONFERENCES,getData());
-    }
-
-    public void goToAttend(ActionEvent actionEvent){
-        FrasinuApplication.createScreen(Screen.ATTEND,getData());
-    }
-    public void seeConferenceInfo(ActionEvent ac){
-        FrasinuApplication.changeScreen(Screen.CONFERENCEINFO,getData());
-    }
-
+    public void seeConferenceInfo(ActionEvent ac){FrasinuApplication.changeScreen(Screen.CONFERENCEINFO,getData());}
     public void goToAuthor(ActionEvent ac){
+
         Author author=userService.findIfUserIsAuthor(new FindIfUserIsAuthorRequest((int)getData().get("idUser"),(int)getData().get("idEdition")));
         if(author==null) {
             Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
@@ -70,10 +56,10 @@ public class ScheduleController extends BaseController{
             alert.setContentText("Do you want to register?");
 
             Optional<ButtonType> result = alert.showAndWait();
-            if (result.get() == ButtonType.OK){
+            if (result.get() == ButtonType.OK) {
                 showDialog("You are now registered as an author. Please complete your personal info on the left side to " +
                         "complete the registration.", "Info!");
-                author=authorService.addAuthor(new CreateAuthorRequest("", "", userService.findById(new FindByIdRequest((int) getData().get("idUser"))),
+                author = authorService.addAuthor(new CreateAuthorRequest("", "", userService.findById(new FindByIdRequest((int) getData().get("idUser"))),
                         conferenceEditionService.findByConferenceEditionId(new FindByConferenceEditionIdRequest((int) getData().get("idEdition")))));
             } else {
                 return;
@@ -85,10 +71,6 @@ public class ScheduleController extends BaseController{
     }
     public void goToPCMember(ActionEvent ac){
         FrasinuApplication.changeScreen(Screen.PCMEMBER, getData());
-    }
-
-    public void goToSteeringCom(ActionEvent ac){
-        FrasinuApplication.changeScreen(Screen.STEERING, getData());
     }
 
 }
